requirements.txt
PyYAML~=5.3.1
discord.py